Use local dummy server to test behavior of nonexistent file
